Template Importer Basics - front-end
